requirements.txt
praw
tqdm
python-dotenv
httpx